+++ CHANGELOG.md
@@ -8,3 +8,5 @@ Version 2.0.0
 Breaking Changes
 ----------------
 - Node.js versions below 10 are no longer supported.
+- lib/authenticate no longer requires config to be passed to it.  Breaks
+  playerManager.